Updated MVC skeleton

+++ src/main/java/pl/sda/transporeon/currencyexchange/repository/ExchangeRateRepository.java
@@ -3,5 +3,5 @@ package pl.sda.transporeon.currencyexchange.repository;
 import org.springframework.data.repository.CrudRepository;
 import pl.sda.transporeon.currencyexchange.model.ExchangeRate;
 
-public interface ExchangeRateRepository extends CrudRepository<ExchangeRate, Long> {
+public interface ExchangeRateRepository extends CrudRepository<ExchangeRate, Integer> {
 }
